requirements.txt
customtkinter>=5.2.2
Pillow>=10.0.1
tkcalendar>=1.6.1
plyer>=2.1.0
pyinstaller>=6.0.0
